refactor(policies): adopt shared dtype management across policies
Use config.dtype and PreTrainedPolicy.post_init() to configure policy
parameter precision. Replace model-local precision switches, casting
routines and duplicated checkpoint loaders with the shared policy API.

Declare existing FP32 exceptions with full paths on each policy class.
Preserve the PI-family float32 defaults, load EO1's Qwen backbone in the
requested dtype, and finalize MolmoAct2's base model before its existing
PEFT adapter initialization. Align helper loader arguments and activations
at mixed-dtype boundaries with the receiving parameters.

Move concrete-policy precision tests out of the base API test module, add
coverage for backbone loading and checkpoint round trips, and update model
usage documentation.

+++ src/lerobot/policies/smolvla/smolvlm_with_expert.py
@@ -91,7 +91,7 @@ class SmolVLMWithExpertModel(nn.Module):
             print(f"Loading  {model_id} weights ...")
             self.vlm = AutoModelForImageTextToText.from_pretrained(
                 model_id,
-                torch_dtype="bfloat16",
+                dtype=torch.float32,
                 low_cpu_mem_usage=True,
             )
             config = self.vlm.config
